【支出列表】列表页更新样式，详情页添加父级字段更新

+++ frontend/src/views/OutlayCat/List.vue
@@ -1,48 +1,46 @@
 <template>
-  <!-- 列表页
-  <div v-for="item in items">
-    <div>
-      {{ item.name }}
-    </div>
-    <ElButton type="primary" @click="create(item)">添加</ElButton>
-    <ElButton type="primary" @click="update(item)">编辑</ElButton>
-    <ElButton type="warning" @click="remove(item)">删除</ElButton>
-  </div> -->
-  <ElTree lazy :load="load" :props="treeProps" node-key="id">
-    <template #default="{ data: item }: { node: Node, data: OutlayCat }">
+  <ElTree :load="load" :props="treeProps" :default-expanded-keys="[ROOT_ID]" node-key="id" lazy>
+    <template #default="{ node, data: item }: { node: Node, data: OutlayCat }">
       <label>{{ item.name }}</label>
-      <ElButton type="success" @click="create(item)" :icon="Plus" size="small" />
-      <ElButton type="primary" @click="update(item)" :icon="Edit" size="small" />
-      <ElButton type="danger" @click="remove(item)" :icon="Delete" size="small" />
+      <ElButton type="success" @click="create(item)" :icon="Plus" circle />
+      <template v-if="node.level > 1">
+        <ElButton type="primary" @click="update(item)" :icon="Edit" circle />
+        <ElButton type="danger" @click="remove(item)" :icon="Delete" circle />
+      </template>
     </template>
   </ElTree>
 </template>
 <script setup lang="ts">
 import { reactive } from 'vue';
-import { onBeforeRouteUpdate, useRoute, useRouter } from 'vue-router';
-import { OutlayCat, IListDto as OutlayCatListDto } from './model';
-import { ElTree, ElButton } from 'element-plus';
+import { useRoute, useRouter } from 'vue-router';
+import { ElTree, ElButton, ElMessage, ElMessageBox } from 'element-plus';
 import { Plus, Edit, Delete } from '@element-plus/icons-vue';
-import 'element-plus/es/components/tree/style/css';
-import 'element-plus/es/components/button/style/css';
 import {
   LoadFunction,
   TreeData,
   TreeOptionProps,
 } from 'element-plus/es/components/tree/src/tree.type.mjs';
 import Node from 'element-plus/es/components/tree/src/model/node';
-const route = useRoute();
+import 'element-plus/es/components/tree/style/css';
+import 'element-plus/es/components/button/style/css';
+import { OutlayCat, ROOT, ROOT_ID } from './model';
+// const route = useRoute();
 const router = useRouter();
-const dto = reactive<OutlayCatListDto>({ parentId: null });
-const items = reactive<OutlayCat[]>([]);
 const treeProps: TreeOptionProps = {
   label: 'name',
   isLeaf: (data, _node) => !(data as OutlayCat).hasChildren,
 };
 const load: LoadFunction = async (node: Node, resolve: (data: TreeData) => void) => {
-  const parentId = node.level === 0 ? null : (node.data as OutlayCat).id;
-  const items = await OutlayCat.list({ parentId });
-  resolve(items);
+  switch (node.level) {
+    case 0:
+      resolve([ROOT]);
+      break;
+    default:
+      const parentId = (node.data as OutlayCat).id;
+      const items = await OutlayCat.list({ parentId });
+      resolve(items);
+      break;
+  }
 };
 const create = (item: OutlayCat) => {
   router.push({ path: 'detail', query: { mode: 'create', parentId: item.id } });
@@ -51,18 +49,22 @@ const update = (item: OutlayCat) => {
   router.push({ path: 'detail', query: { id: item.id, mode: 'update' } });
 };
 const remove = async (item: OutlayCat) => {
-  await item.delete();
-  const i = items.indexOf(item);
-  items.splice(i, 1);
+  const confirmed = await ElMessageBox.confirm(`确认删除${item.name}？`, '', {
+    confirmButtonText: '确认',
+    cancelButtonText: '取消',
+    type: 'warning',
+  })
+    .then(() => true)
+    .catch(() => false);
+  if (confirmed) {
+    const res = await item.delete();
+    if (res) {
+      ElMessage.success('删除成功');
+    } else {
+      ElMessage.error('删除失败');
+    }
+  }
 };
-const init = async () => {
-  items.splice(0);
-  const list = await OutlayCat.list(dto);
-  items.push(...list);
-};
+const init = async () => {};
 init();
-// onBeforeRouteUpdate(async (to, from) => {
-//   console.log(to, from);
-//   await init();
-// });
 </script>
